requirements.txt
# requirements.txt
psutil
psycopg2-binary
pandas
pynvml
requests
py-cpuinfo
wmi